Use consistent uint256 types

+++ contracts/core/GovernanceController.sol
@@ -17,13 +17,13 @@ contract GovernanceController is Ownable {
     Role[] private roles;
 
     event CallFunction(address indexed sender, address indexed contractAddress);
-    event AddRole(uint indexed roleId, string name, address[] members, address[] contracts, bytes4[][] functionSelectors);
-    event GrantRole(uint indexed roleId, address indexed account);
-    event RevokeRole(uint indexed roleId, address indexed account);
-    event AddRoleFunctions(uint indexed roleId, address[] contracts, bytes4[][] functionSelectors);
-    event RemoveRoleFunctions(uint indexed roleId, address[] contracts, bytes4[][] functionSelectors);
+    event AddRole(uint256 indexed roleId, string name, address[] members, address[] contracts, bytes4[][] functionSelectors);
+    event GrantRole(uint256 indexed roleId, address indexed account);
+    event RevokeRole(uint256 indexed roleId, address indexed account);
+    event AddRoleFunctions(uint256 indexed roleId, address[] contracts, bytes4[][] functionSelectors);
+    event RemoveRoleFunctions(uint256 indexed roleId, address[] contracts, bytes4[][] functionSelectors);
 
-    modifier roleExists(uint _roleId) {
+    modifier roleExists(uint256 _roleId) {
         require(_roleId < roles.length, "Role does not exist");
         _;
     }
@@ -34,7 +34,7 @@ contract GovernanceController is Ownable {
      **/
     function getRoles() external view returns (string[] memory roleNames) {
         roleNames = new string[](roles.length);
-        for (uint i = 0; i < roles.length; i++) {
+        for (uint256 i = 0; i < roles.length; i++) {
             roleNames[i] = roles[i].name;
         }
     }
@@ -45,7 +45,7 @@ contract GovernanceController is Ownable {
      * @param _account address of account
      * @return hasRole whether or not account holds tole
      **/
-    function hasRole(uint _roleId, address _account) external view returns (bool) {
+    function hasRole(uint256 _roleId, address _account) external view returns (bool) {
         return roles[_roleId].members[_account];
     }
 
@@ -57,7 +57,7 @@ contract GovernanceController is Ownable {
      * @return isAuthorized whether or not role is authorized to call function
      **/
     function hasFunction(
-        uint _roleId,
+        uint256 _roleId,
         address _contract,
         bytes4 _functionSelector
     ) external view returns (bool) {
@@ -71,7 +71,7 @@ contract GovernanceController is Ownable {
      * @param _encodedCalldata encoded function call
      **/
     function callFunction(
-        uint _roleId,
+        uint256 _roleId,
         address _contract,
         bytes calldata _encodedCalldata
     ) external roleExists(_roleId) {
@@ -105,13 +105,13 @@ contract GovernanceController is Ownable {
         Role storage role = roles.push();
         role.name = _name;
 
-        for (uint i = 0; i < _members.length; i++) {
+        for (uint256 i = 0; i < _members.length; i++) {
             role.members[_members[i]] = true;
         }
 
-        for (uint i = 0; i < _contracts.length; i++) {
+        for (uint256 i = 0; i < _contracts.length; i++) {
             address contractAddress = _contracts[i];
-            for (uint j = 0; j < _functionSelectors[i].length; j++) {
+            for (uint256 j = 0; j < _functionSelectors[i].length; j++) {
                 role.functions[_getFunctionId(contractAddress, _functionSelectors[i][j])] = true;
             }
         }
@@ -124,7 +124,7 @@ contract GovernanceController is Ownable {
      * @param _roleId id of role to grant
      * @param _account address to grant role to
      **/
-    function grantRole(uint _roleId, address _account) external onlyOwner roleExists(_roleId) {
+    function grantRole(uint256 _roleId, address _account) external onlyOwner roleExists(_roleId) {
         require(!roles[_roleId].members[_account], "Account already holds role");
         roles[_roleId].members[_account] = true;
         emit GrantRole(_roleId, _account);
@@ -135,7 +135,7 @@ contract GovernanceController is Ownable {
      * @param _roleId id of role to revoke
      * @param _account address to revoke role from
      **/
-    function revokeRole(uint _roleId, address _account) external onlyOwner roleExists(_roleId) {
+    function revokeRole(uint256 _roleId, address _account) external onlyOwner roleExists(_roleId) {
         _revokeRole(_roleId, _account);
     }
 
@@ -143,7 +143,7 @@ contract GovernanceController is Ownable {
      * @notice renounces a role from the sender
      * @param _roleId id of role to renounce
      **/
-    function renounceRole(uint _roleId) external roleExists(_roleId) {
+    function renounceRole(uint256 _roleId) external roleExists(_roleId) {
         _revokeRole(_roleId, msg.sender);
     }
 
@@ -154,14 +154,14 @@ contract GovernanceController is Ownable {
      * @param _functionSelectors list of function selectors to add for each contract
      **/
     function addRoleFunctions(
-        uint _roleId,
+        uint256 _roleId,
         address[] calldata _contracts,
         bytes4[][] calldata _functionSelectors
     ) external onlyOwner roleExists(_roleId) {
-        for (uint i = 0; i < _contracts.length; i++) {
+        for (uint256 i = 0; i < _contracts.length; i++) {
             address contractAddress = _contracts[i];
 
-            for (uint j = 0; j < _functionSelectors[i].length; j++) {
+            for (uint256 j = 0; j < _functionSelectors[i].length; j++) {
                 bytes32 functionId = _getFunctionId(contractAddress, _functionSelectors[i][j]);
                 require(!roles[_roleId].functions[functionId], "Function is already part of role");
                 roles[_roleId].functions[functionId] = true;
@@ -178,14 +178,14 @@ contract GovernanceController is Ownable {
      * @param _functionSelectors list of function selectors to remove for each contract
      **/
     function removeRoleFunctions(
-        uint _roleId,
+        uint256 _roleId,
         address[] calldata _contracts,
         bytes4[][] calldata _functionSelectors
     ) external onlyOwner roleExists(_roleId) {
-        for (uint i = 0; i < _contracts.length; i++) {
+        for (uint256 i = 0; i < _contracts.length; i++) {
             address contractAddress = _contracts[i];
 
-            for (uint j = 0; j < _functionSelectors[i].length; j++) {
+            for (uint256 j = 0; j < _functionSelectors[i].length; j++) {
                 bytes32 functionId = _getFunctionId(contractAddress, _functionSelectors[i][j]);
                 require(roles[_roleId].functions[functionId], "Function is not part of role");
                 roles[_roleId].functions[functionId] = false;
@@ -200,7 +200,7 @@ contract GovernanceController is Ownable {
      * @param _roleId id of role to revoke
      * @param _account address to revoke role from
      **/
-    function _revokeRole(uint _roleId, address _account) private {
+    function _revokeRole(uint256 _roleId, address _account) private {
         require(roles[_roleId].members[_account], "Account does not hold role");
         roles[_roleId].members[_account] = false;
         emit RevokeRole(_roleId, _account);
